refactor: split the many functions in multiple files
