+++ cloud-functions/functions/tsconfig.json
@@ -1,16 +1,17 @@
 {
   "compilerOptions": {
-    "module": "es2022",
+    "module": "nodenext",
     "moduleResolution": "nodenext",
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "outDir": "lib",
     "sourceMap": true,
     "strict": true,
-    "target": "esnext"
+    "target": "esnext",
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true
   },
   "compileOnSave": true,
-  "include": [
-    "src"
-  ]
+  "include": [ "src" ]
 }
